songview refactor - separate the posts and the song into two components with different database calls
